Category DTO implemented.

+++ backend/springJwt/src/main/java/com/PracticaVara/springJwt/service/AnnouncementServices/CategoryService.java
@@ -1,9 +1,14 @@
 package com.PracticaVara.springJwt.service.AnnouncementServices;
 
+import com.PracticaVara.springJwt.DTOs.CategoryDTO;
+import com.PracticaVara.springJwt.model.APIMessage;
 import com.PracticaVara.springJwt.model.Category;
 import com.PracticaVara.springJwt.repository.CategoryRepository;
 import jakarta.servlet.ServletContext;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -41,9 +46,11 @@ public class CategoryService {
         return categoryRepository.findById(id);
     }
 
-    public Category addCategory(Category category, MultipartFile file) throws IOException {
+    public ResponseEntity<?> addCategory(CategoryDTO categoryDTO, MultipartFile file) throws IOException {
+        Category category = new Category();
         if (!file.isEmpty()) {
-            String imageName = category.getName().replaceAll("[^a-zA-Z0-9]", "-").toLowerCase() + ".png";
+
+            String imageName = categoryDTO.getName().replaceAll("[^a-zA-Z0-9]", "-").toLowerCase() + ".png";
 //            File dest = new File(uploadDir + imageName);
 //            file.transferTo(dest);
 //            category.setIconUrl(imageName);
@@ -51,34 +58,53 @@ public class CategoryService {
             Files.copy(file.getInputStream(), destinationFile);
             category.setIconUrl(imageName);
             category.setSearchLink(category.getName().replaceAll("[^a-zA-Z0-9]", "-").toLowerCase());
+            categoryRepository.save(category);
+            return ResponseEntity.status(HttpStatus.CREATED).body(category);
+        } else {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new APIMessage(HttpStatus.BAD_REQUEST, "Fisierul nu a fost gasit"));
         }
-        return categoryRepository.save(category);
+
     }
 
-    public Category updateCategory(Long id, Category categoryDetails, MultipartFile file) throws IOException {
-        Category category = categoryRepository.findById(id).orElseThrow(() -> new RuntimeException("Category not found"));
+    public ResponseEntity<?> updateCategory(Long id, CategoryDTO categoryDetailsDTO, MultipartFile file) throws IOException {
+        Optional<Category> categoryOptional = categoryRepository.findById(id);
+        if(categoryOptional.isPresent()) {
+            Category category = categoryOptional.get();
+            category.setName(categoryDetailsDTO.getName());
+            category.setSearchLink(category.getSearchLink());
 
-        category.setName(categoryDetails.getName());
-        category.setSearchLink(categoryDetails.getSearchLink());
+            if (!file.isEmpty()) {
+                String imageName = category.getName().replaceAll("[^a-zA-Z0-9]", "-").toLowerCase() + ".png";
+                File dest = new File(uploadDir + imageName);
+                file.transferTo(dest);
+                category.setIconUrl(imageName);
+            } else {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new APIMessage(HttpStatus.NOT_FOUND, "Imaginea nu a fost gasita."));
+            }
 
-        if (!file.isEmpty()) {
-            String imageName = category.getName().replaceAll("[^a-zA-Z0-9]", "-").toLowerCase() + ".png";
-            File dest = new File(uploadDir + imageName);
-            file.transferTo(dest);
-            category.setIconUrl(imageName);
+            categoryRepository.save(category);
+            return ResponseEntity.status(HttpStatus.OK).body(category);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new APIMessage(HttpStatus.NOT_FOUND, "Categoria nu a fost gasita."));
         }
 
-        return categoryRepository.save(category);
     }
 
-    public void deleteCategory(Long id) {
-        Category category = categoryRepository.findById(id).orElseThrow(() -> new RuntimeException("Category not found"));
+    public ResponseEntity<?> deleteCategory(Long id) {
+        Optional<Category> categoryOptional = categoryRepository.findById(id);
+        if(categoryOptional.isPresent()) {
+            Category category = categoryOptional.get();
+            File file = new File(uploadDir + category.getIconUrl());
+            if (file.exists()) {
+                file.delete();
+                categoryRepository.delete(category);
+                return ResponseEntity.status(HttpStatus.OK).body(new APIMessage(HttpStatus.OK, "Categoria a fost stearsa cu succes."));
+            } else {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new APIMessage(HttpStatus.NOT_FOUND, "Fisierele nu au fost gasite."));
+            }
 
-        File file = new File(uploadDir + category.getIconUrl());
-        if (file.exists()) {
-            file.delete();
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new APIMessage(HttpStatus.NOT_FOUND, "Categoria nu a fost gasita."));
         }
-
-        categoryRepository.delete(category);
     }
 }
